Added Chunked Messages
This code is currently not optimized. The memory should not be reallocated each time, instead, the old ByteBuffer should be reused. When receiving, only the most necessary information should be stored so that less memory is used. Every chunk that was successfully sent or read should be cleared immediately to free up memory.